Rename ipywidgets message passing
