+++ src/datascience-ui/ipywidgets/common/incompatibleWidgetHandler.ts
@@ -4,7 +4,7 @@
 'use strict';
 
 import * as semver from 'semver';
-import { WidgetScriptSource } from '../../../kernels/ipywidgets/types';
+import { WidgetScriptSource } from '../../../kernels/ipywidgets-message-coordination/types';
 const supportedVersionOfQgrid = '1.1.1';
 const qgridModuleName = 'qgrid';
