+++ src/kernels/ipywidgets-message-coordination/cdnWidgetScriptSourceProvider.ts
