+++ src/kernels/ipywidgets-message-coordination/ipyWidgetScriptSourceProvider.ts
