+++ src/kernels/ipywidgets-message-coordination/localWidgetScriptSourceProvider.ts
