+++ src/kernels/ipywidgets-message-coordination/remoteWidgetScriptSourceProvider.ts
